WIP: Start adding more tests.
