+++ tests/utils.ts
@@ -1,3 +1,5 @@
+/* eslint-disable @typescript-eslint/no-floating-promises */
+
 import type { Express } from "express";
 import type { Server } from "http";
 import type { Test } from "supertest";
@@ -9,6 +11,7 @@ import { createApp } from "../src/server";
 import { Class, Student } from "../src/models";
 import { APIKey } from "../src/models/api_key";
 import { config } from "dotenv";
+import { v4 } from "uuid";
 import { getDatabaseConnection } from "../src/database";
 import { createConnection, Connection } from "mysql2/promise";
 import { hashAPIKey } from "../src/authorization";
@@ -93,3 +96,58 @@ export function createTestApp(db: Sequelize): Express {
 export function runApp(app: Express, port = 8080, callback?: () => void): Server {
   return app.listen(port, callback);
 }
+
+export async function setupStudentInClasses() {
+  const educator = await Educator.create({
+    first_name: v4(),
+    last_name: v4(),
+    password: v4(),
+    email: v4(),
+    verified: 1,
+    verification_code: v4(),
+    username: v4(),
+  });
+  const class1 = await Class.create({
+    name: v4(),
+    educator_id: educator.id,
+    code: v4(),
+    expected_size: 1,
+  });
+  const class2 = await Class.create({
+    name: v4(),
+    educator_id: educator.id,
+    code: v4(),
+    expected_size: 1,
+  });
+  const student = await Student.create({
+    email: v4(),
+    username: v4(),
+    password: v4(),
+    verification_code: class1.code,
+    verified: 0,
+  });
+
+  const sc1 = await StudentsClasses.create({
+    student_id: student.id,
+    class_id: class1.id,
+  });
+  const sc2 = await StudentsClasses.create({
+    student_id: student.id,
+    class_id: class2.id,
+  });
+
+  const cleanup = async () => {
+    // Sometimes we want to remove the class-student associations during the test
+    // In which case we can check whether it exists first
+    (await StudentsClasses.findOne({ where: { student_id: student.id, class_id: class1.id } }))?.destroy();
+    (await StudentsClasses.findOne({ where: { student_id: student.id, class_id: class2.id } }))?.destroy();
+    await class1.destroy();
+    await class2.destroy();
+    await educator.destroy();
+    await student.destroy();
+  };
+
+  return { student, educator, class1, class2, sc1, sc2, cleanup };
+}
+
+
